Add files via upload
This should be able to push tasks up to server and down to client, however I ran into issues with my cs account where I can't do anything on the server. I'm working on resolving this, but I probably won't be able to until Monday.
I'm leaving the release up to you Cass because you are more familiar with the ui and such on here, and also so you can add your stuff.